improve transaction locking in sqlite dbs
