+++ app/background-process/dbs/history.js
@@ -5,10 +5,9 @@ import url from 'url'
 import zerr from 'zerr'
 import multicb from 'multicb'
 import rpc from 'pauls-electron-rpc'
-import FnQueue from 'function-queue'
 import manifest from '../api-manifests/history'
 import { cbPromise } from '../../lib/functions'
-import { setupDatabase2 } from '../../lib/bg/sqlite-tools'
+import { setupSqliteDB, makeSqliteTransactor } from '../../lib/bg/db'
 
 const BadParam = zerr('BadParam', '% must be a %')
 const InvalidCmd = zerr('InvalidCommand', '% is not a valid command')
@@ -18,7 +17,7 @@ const InvalidCmd = zerr('InvalidCommand', '% is not a valid command')
 var db
 var migrations
 var setupPromise
-var addVisitQueue = FnQueue()
+var tx
 
 // exported methods
 // =
@@ -27,14 +26,15 @@ export function setup () {
   // open database
   var dbPath = path.join(app.getPath('userData'), 'History')
   db = new sqlite3.Database(dbPath)
-  setupPromise = setupDatabase2(db, migrations, '[HISTORY]')
+  setupPromise = setupSqliteDB(db, migrations, '[HISTORY]')
+  tx = makeSqliteTransactor(setupPromise)
 
   // wire up RPC
   rpc.exportAPI('beakerHistory', manifest, { addVisit, getVisitHistory, getMostVisited, search, removeVisit, removeAllVisits, removeVisitsAfter })
 }
 
 export function addVisit ({url, title}) {
-  return setupPromise.then(v => cbPromise(cb => {
+  return tx(cb => {
     // validate parameters
     cb = cb || (()=>{})
     if (!url || typeof url != 'string')
@@ -43,61 +43,55 @@ export function addVisit ({url, title}) {
       return cb(new BadParam('title', 'string'))
 
     // get current stats
-    addVisitQueue.push(endTransaction => {
-      db.serialize(() => {
-        db.run('BEGIN TRANSACTION;')
-        db.get('SELECT * FROM visit_stats WHERE url = ?;', [url], (err, stats) => {
+    db.serialize(() => {
+      db.run('BEGIN TRANSACTION;')
+      db.get('SELECT * FROM visit_stats WHERE url = ?;', [url], (err, stats) => {
+        if (err) return cb(err)
+
+        var done = multicb()
+        var ts = Date.now()
+
+        // visited within 1 hour?
+        db.get('SELECT rowid, * from visits WHERE url = ? AND ts > ? ORDER BY ts DESC LIMIT 1', [url, ts - 1000 * 60 * 60], (err, visit) => {
           if (err) return cb(err)
 
-          var done = multicb()
-          var ts = Date.now()
+          db.serialize(() => {
+            if (!stats) {
+              // yes, create new stat and search entries
+              db.run('INSERT INTO visit_stats (url, num_visits, last_visit_ts) VALUES (?, ?, ?);', [url, 1, ts], done())
+              db.run('INSERT INTO visit_fts (url, title) VALUES (?, ?);', [url, title], done())
+            } else {
+              // no, update stats
+              var num_visits = (+stats.num_visits||1) + 1
+              db.run('UPDATE visit_stats SET num_visits = ?, last_visit_ts = ? WHERE url = ?;', [num_visits, ts, url], done())
+            }
 
-          // visited within 1 hour?
-          db.get('SELECT rowid, * from visits WHERE url = ? AND ts > ? ORDER BY ts DESC LIMIT 1', [url, ts - 1000 * 60 * 60], (err, visit) => {
-            if (err) return cb(err)
-
-            db.serialize(() => {
-              if (!stats) {
-                // yes, create new stat and search entries
-                db.run('INSERT INTO visit_stats (url, num_visits, last_visit_ts) VALUES (?, ?, ?);', [url, 1, ts], done())
-                db.run('INSERT INTO visit_fts (url, title) VALUES (?, ?);', [url, title], done())
-              } else {
-                // no, update stats
-                var num_visits = (+stats.num_visits||1) + 1
-                db.run('UPDATE visit_stats SET num_visits = ?, last_visit_ts = ? WHERE url = ?;', [num_visits, ts, url], done())
-              }
-
-              if (visit) {
-                // update visit ts and title
-                db.run('UPDATE visits SET ts = ?, title = ? WHERE rowid = ?', [ts, title, visit.rowid], done())
-              } else {
-                // log visit
-                db.run('INSERT INTO visits (url, title, ts) VALUES (?, ?, ?);', [url, title, ts], done())
-              }
-              db.run('COMMIT;', done())
-            })
-          })
-
-          done(err => {
-            endTransaction()
-            cb(err)
+            if (visit) {
+              // update visit ts and title
+              db.run('UPDATE visits SET ts = ?, title = ? WHERE rowid = ?', [ts, title, visit.rowid], done())
+            } else {
+              // log visit
+              db.run('INSERT INTO visits (url, title, ts) VALUES (?, ?, ?);', [url, title, ts], done())
+            }
+            db.run('COMMIT;', done())
           })
         })
+        done(cb)
       })
     })
-  }))
+  })
 }
 
 export function getVisitHistory ({ offset, limit }) {
-  return setupPromise.then(v => cbPromise(cb => {
+  return tx(cb => {
     offset = offset || 0
     limit = limit || 50
     db.all('SELECT * FROM visits ORDER BY ts DESC LIMIT ? OFFSET ?', [limit, offset], cb)
-  }))
+  })
 }
 
 export function getMostVisited ({ offset, limit }) {
-  return setupPromise.then(v => cbPromise(cb => {
+  return tx(cb => {
     offset = offset || 0
     limit = limit || 50
     db.all(`
@@ -109,11 +103,11 @@ export function getMostVisited ({ offset, limit }) {
         ORDER BY num_visits DESC, last_visit_ts DESC
         LIMIT ? OFFSET ?
     `, [limit, offset], cb)
-  }))
+  })
 }
 
 export function search (q) {
-  return setupPromise.then(v => cbPromise(cb => {
+  return tx(cb => {
     if (!q || typeof q != 'string')
       return cb(new BadParam('q', 'string'))
 
@@ -132,11 +126,11 @@ export function search (q) {
         ORDER BY visit_stats.num_visits DESC
         LIMIT 10;
     `, [q], cb)
-  }))
+  })
 }
 
 export function removeVisit (url) {
-  return setupPromise.then(v => cbPromise(cb => {
+  return tx(cb => {
     // validate parameters
     cb = cb || (()=>{})
     if (!url || typeof url != 'string')
@@ -149,28 +143,28 @@ export function removeVisit (url) {
       db.run('DELETE FROM visit_fts WHERE url = ?;', url)
       db.run('COMMIT;', cb)
     })
-  }))
+  })
 }
 
 export function removeVisitsAfter (timestamp) {
-  return setupPromise.then(v => cbPromise(cb => {
+  return tx(cb => {
     cb = cb || (()=>{})
 
     db.run('BEGIN TRANSACTION;')
     db.run('DELETE FROM visits WHERE ts >= ?;', timestamp)
     db.run('DELETE FROM visit_stats WHERE last_visit_ts >= ?;', timestamp)
     db.run('COMMIT;', cb)
-  }))
+  })
 }
 
 export function removeAllVisits () {
-  return setupPromise.then(v => cbPromise(cb => {
+  return tx(cb => {
     cb = cb || (()=>{})
 
     db.run('DELETE FROM visits;')
     db.run('DELETE FROM visit_stats;')
     db.run('DELETE FROM visit_fts;')
-  }))
+  })
 }
 
 // internal methods
